requirements.txt
google-cloud-speech
google-cloud-storage
python-docx
pydub
